fix based on the comments
