--- qiskit_optimization/applications/ising/__init__.py
@@ -1,42 +1,0 @@
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2019, 2021.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-"""
-Ising Models (:mod:`qiskit_optimization.applications.ising`)
-============================================================
-
-Ising models for optimization application problems
-
-.. currentmodule:: qiskit_optimization.applications.ising
-
-Ising Models
-============
-
-.. autosummary::
-   :toctree: ../stubs/
-   :nosignatures:
-
-   base_optimization_application
-   clique
-   exact_cover
-   graph_optimization_application
-   graph_partition
-   knapsack
-   max_cut
-   number_partition
-   set_packing
-   stable_set
-   tsp
-   vehicle_routing
-   vertex_cover
-
-"""
