+++ qiskit_optimization/applications/vertex_cover.py
@@ -10,7 +10,7 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""An application class for the graph partitioning."""
+"""An application class for the vertex cover."""
 
 from typing import Dict, List, Optional
 
@@ -23,53 +23,48 @@ from qiskit_optimization.problems.quadratic_program import QuadraticProgram
 from .graph_optimization_application import GraphOptimizationApplication
 
 
-class GraphPartition(GraphOptimizationApplication):
-    """Convert a graph partition [1] instance based on a graph of NetworkX into a
-    :class:`~qiskit_optimization.problems.QuadraticProgram`
+class VertexCover(GraphOptimizationApplication):
+    """Optimization application for the "vertex cover" [1] problem based on a NetworkX graph.
 
     References:
-        [1]: "Graph partition", https://en.wikipedia.org/wiki/Graph_partition
+        [1]: "Vertex cover", https://en.wikipedia.org/wiki/Vertex_cover
     """
 
     def to_quadratic_program(self) -> QuadraticProgram:
-        """Convert a graph partition instance into a
+        """Convert a vertex cover instance into a
         :class:`~qiskit_optimization.problems.QuadraticProgram`
 
         Returns:
             The :class:`~qiskit_optimization.problems.QuadraticProgram` created
-            from the graph partition instance.
+            from the vertex cover instance.
         """
-        mdl = Model(name='Graph partinion')
+        mdl = Model(name='Vertex cover')
         n = self._graph.number_of_nodes()
         x = {i: mdl.binary_var(name='x_{0}'.format(i)) for i in range(n)}
+        objective = mdl.sum(x[i] for i in x)
         for w, v in self._graph.edges:
-            self._graph.edges[w, v].setdefault('weight', 1)
-        objective = mdl.sum(self._graph.edges[i, j]['weight'] *
-                            (x[i] + x[j] - 2*x[i]*x[j]) for i, j in self._graph.edges)
+            mdl.add_constraint(x[w] + x[v] >= 1)
         mdl.minimize(objective)
-        mdl.add_constraint(mdl.sum([x[i] for i in x]) == n//2)
         op = QuadraticProgram()
         op.from_docplex(mdl)
         return op
 
-    def interpret(self, result: OptimizationResult) -> List[List[int]]:
+    def interpret(self, result: OptimizationResult) -> List[int]:
         """Interpret a result as a list of node indices
 
         Args:
             result : The calculated result of the problem
 
         Returns:
-            A list of node indices divided into two groups.
+            A list of node indices whose corresponding variable is 1
         """
-        partition = [[], []]  # type: List[List[int]]
+        vertex_cover = []
         for i, value in enumerate(result.x):
-            if value == 0:
-                partition[0].append(i)
-            else:
-                partition[1].append(i)
-        return partition
+            if value:
+                vertex_cover.append(i)
+        return vertex_cover
 
-    def draw_graph(self, result: Optional[OptimizationResult] = None,
+    def draw(self, result: Optional[OptimizationResult] = None,
                    pos: Optional[Dict[int, np.ndarray]] = None) -> None:
         """Draw a graph with the result. When the result is None, draw an original graph without
         colors.
@@ -81,10 +76,10 @@ class GraphPartition(GraphOptimizationApplication):
         if result is None:
             nx.draw(self._graph, pos=pos, with_labels=True)
         else:
-            nx.draw(self._graph, node_color=self._node_colors, pos=pos, with_labels=True)
+            nx.draw(self._graph, node_color=self._node_colors(result), pos=pos, with_labels=True)
 
     def _node_colors(self, result: OptimizationResult) -> List[str]:
-        # Return a list of strings for draw_graph.
+        # Return a list of strings for draw.
         # Color a node with red when the corresponding variable is 1.
-        # Otherwise color it with blue.
-        return ['r' if value else 'b' for value in result.x]
+        # Otherwise color it with darkgrey.
+        return ['r' if value else 'darkgrey' for value in result.x]
